pyproject.toml
[project]
name = "depenses-eclairees"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.13"
dependencies = [
    "docx2txt",
    "python-dotenv",
    "faiss-cpu",
    "ipykernel",
    "jupyter",
    "olefile",
    "openai",
    "openpyxl",
    "pdf2image",
    "pandas",
    "Pillow",
    "psycopg[binary]",
    "PyPDF2",
    "PyCryptodome",
    "pymupdf",
    "pytesseract",
    "python-magic",
    "scikit-learn",
    "tiktoken",
    "tqdm",

    # Web app
    "streamlit",
    "Authlib",
    "sqlalchemy (>=2.0.43,<3.0.0)",

    # Django
    "Django~=5.2",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-dsfr (>=3.2.0,<4.0.0)",
    "mozilla-django-oidc (>=4.0.1,<5.0.0)",
    "django-lasuite (>=0.0.14,<0.0.15)",
    "django-magicauth (>=1.0.2,<2.0.0)",
]

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "freezegun",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "pytest-factoryboy",
    "pytest-responses",
    "pytest-ruff",
    "ruff",
]


[tool.ruff]
line-length = 120
exclude = ["**/migrations/*.py"]

[tool.ruff.lint]
select = [
    "I",  # isort
    "F",  # pyflakes
    "E",  # pycodestyle
    "DJ",  # flake8-django
]

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "django",
    "third-party",
    "pandas",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
# Group all Django and Pandas imports into a separate section.
django = ["django"]
pandas = ["pandas", "numpy"]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "docia.settings"
